requirements.txt
boto3~=1.28.28
python-dotenv==1.0.0
